requirements.txt
# Database
mysql-connector-python>=8.0.0
tabulate>=0.9.0

# Google Cloud and AI
google-adk>=0.3.0  # For Google Agent Development Kit

# Environment and Utilities
python-dotenv>=1.0.0  # For environment variables 